refactor: Удалить функционал, связанный с веткой add-feed
