--- src/main/java/ru/yandex/practicum/filmorate/model/event/EventOperation.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.model.event;
-
-public enum EventOperation {
-    ADD,
-    UPDATE,
-    REMOVE
-}
