--- src/main/java/ru/yandex/practicum/filmorate/model/event/EventType.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.model.event;
-
-public enum EventType {
-    LIKE,
-    REVIEW,
-    FRIEND
-}
