+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.model.event.Event;
 
 import java.util.Collection;
 
@@ -25,6 +24,4 @@ public interface UserStorage {
     void deleteFriend(Integer id, Integer friendId);
 
     Collection<User> getCommonFriends(Integer id, Integer otherId);
-
-    Collection<Event> getEventFeed(Integer userId);
 }
