+++ src/main/resources/schema.sql
@@ -67,13 +67,3 @@ CREATE TABLE IF NOT EXISTS FRIENDS
     FRIEND_ID       INTEGER NOT NULL REFERENCES USERS (USER_ID) ON DELETE CASCADE,
     CONSTRAINT FRIENDS_PK PRIMARY KEY (USER_ID, FRIEND_ID)
 );
-
-CREATE TABLE IF NOT EXISTS FEEDS
-(
-    EVENT_ID INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    EVENT_TYPE VARCHAR(10) NOT NULL,
-    OPERATION VARCHAR(10) NOT NULL,
-    USER_ID INTEGER REFERENCES USERS (USER_ID) ON DELETE CASCADE,
-    ENTITY_ID INTEGER,
-    EVENT_TIMESTAMP TIMESTAMP
-);
